Don't generate locations in po/grip-2.2.po by default
The file:linenum notation in po/grip-2.2.po increase the likelihood of
merge conflicts.  Leave them out by default.

As they're useful for anyone wanting to work on the translations,
allow them the be re-enabled by

   ./configure --enable-gettext-location

Which defaults to off.

The setting in configure.in will be propogated to configure by
autoconf.  My autoconf is a different version than the one used to
generate the current configure, so I've worked the resulting changes
into configure by hand.